toolkit/mozapps/downloads/tests/chrome/test_clear_button_disabled.xul

<?xml version="1.0"?>
<!-- This Source Code Form is subject to the terms of the Mozilla Public
   - License, v. 2.0. If a copy of the MPL was not distributed with this
   - file, You can obtain one at http://mozilla.org/MPL/2.0/.  -->
<!--
 * This tests 437422.  This test basically intends to checks if the clear list
 * button is disabled when:
 * 1.  an invalid search string has been entered into the search box.
 * 2.  active downloads are present in the dm ui
 * 3.  we have both case (1) and (2)
-->

<window title="Download Manager Test"
        onload="runTest();"
        xmlns:html="http://www.w3.org/1999/xhtml"
        xmlns="http://www.mozilla.org/keymaster/gatekeeper/there.is.only.xul">

  <script type="application/javascript"
          src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"/>
  <script type="application/javascript"
          src="utils.js"/>

  <script type="application/javascript">

  <![CDATA[

let ios = Cc["@mozilla.org/network/io-service;1"].getService(Ci.nsIIOService);
let dmFile = Cc["@mozilla.org/file/directory_service;1"].
             getService(Ci.nsIProperties).get("TmpD", Ci.nsIFile);
dmFile.append("dm-ui-test.file");
dmFile.createUnique(Ci.nsIFile.NORMAL_FILE_TYPE, 0666);
let gTestPath = ios.newFileURI(dmFile).spec;

const DoneDownloadData = [
  { name: "Dead",
    source: "https://bugzilla.mozilla.org/attachment.cgi?id=266520",
    target: gTestPath,
    startTime: 1180493839859230,
    endTime: 1180493839859239,
    state: Ci.nsIDownloadManager.DOWNLOAD_CANCELED,
    currBytes: 0, maxBytes: -1, preferredAction: 0, autoResume: 0 }
];

const ActiveDownloadData = [
  { name: "Patch",
    source: "https://bugzilla.mozilla.org/attachment.cgi?id=266520",
    target: gTestPath,
    startTime: 1180493839859230,
    endTime: 1180493839859239,
    state: Ci.nsIDownloadManager.DOWNLOAD_DOWNLOADING,
    currBytes: 0, maxBytes: -1, preferredAction: 0, autoResume: 0 }
];

function runTest()
{
  var dmui = getDMUI();
  if (!dmui) {
    todo(false, "skip test for toolkit download manager UI");
    return;
  }

  setCleanState();
  populateDM(DoneDownloadData);

  let obs = Cc["@mozilla.org/observer-service;1"].
            getService(Ci.nsIObserverService);
  const DLMGR_UI_DONE = "download-manager-ui-done";

  let testPhase = 0;
  let testObs = {
    observe: function(aSubject, aTopic, aData) {
      let win = aSubject.QueryInterface(Ci.nsIDOMWindow);
      win.focus();
      let doc = win.document;
      let searchbox = doc.getElementById("searchbox");
      let clearButton = doc.getElementById("clearListButton");

      switch (testPhase++) {
        case 0:
          // Ensure that the clear list button is enabled at first
          ok(!clearButton.disabled,
             "The clear list button is  not disabled initially.");

          // Now, insert an nonsensical search string - nothing should show up,
          // and the button should be disabled in the next test phase
          searchbox.value = "Nonsensical";
          searchbox.doCommand();

          break;
        case 1:
          ok(clearButton.disabled,
             "The clear list button is disabled with a nonsensical search " +
             "term entered");

          // Clear the search box
          searchbox.value = "";
          searchbox.doCommand();
          break;

        case 2:
          // Populate the download manager with an active download now, and
          // rebuild the list
          let dm = Cc["@mozilla.org/download-manager;1"].
                   getService(Ci.nsIDownloadManager);
          populateDM(ActiveDownloadData);
          dm.cleanUp();

          break;
        case 3:
          ok(clearButton.disabled,
             "The clear list button is disabled when we only have an active " +
             "download");

          // Now, insert an nonsensical search string - only the active download
          // should show up, and the button should be disabled in the next test
          // phase
          searchbox.value = "Nonsensical";
          searchbox.doCommand();
          break;
        case 4:
          ok(clearButton.disabled,
             "The clear list button is disabled with a nonsensical search " +
             "term entered and one active download");

          obs.removeObserver(testObs, DLMGR_UI_DONE);
          setCleanState();
          SimpleTest.finish();

          break;
      }
    }
  };

  obs.addObserver(testObs, DLMGR_UI_DONE, false);

  dmui.show();

  SimpleTest.waitForExplicitFinish();
}

  ]]>
  </script>

  <body xmlns="http://www.w3.org/1999/xhtml">
    <p id="display"></p>
    <div id="content" style="display: none"></div>
    <pre id="test"></pre>
  </body>
</window>
